refactor: put demos inside components just like what ng-devui did
