+++ src/app/components/color-picker/color-picker-demo/color-picker-demo.component.html
