+++ src/app/components/fluent-slider/fluent-slider-demo/fluent-slider-demo.component.html
