+++ src/app/components/image-preview/image-preview-demo/image-preview-demo.component.html
