+++ src/app/components/smart-input/smart-input-demo/smart-input-demo.component.ts
